Imported old settings to the code & bruteforce
Added functionalities for the body generation (InputData folder) and bruteforced DataCode so that it now at least works.
the problem is that for some reason, strings dont parse to floats correctly anymore. For example: "3.1415" will be converted to 31415.0f.
The current bruteforce solution is to divide the number read by 10^5, since this returns the same number, due to the output of Octave only writing 5 decimal digits.